% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_rss_pull.R
\name{google_rss_pull}
\alias{google_rss_pull}
\title{Miscellaneous Codes for collecting text from RSS feeds from Google}
\usage{
google_rss_pull(url, group, country)
}
\arguments{
\item{url}{}
}
\description{

}
\examples{

}
