% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_txt_pull.R
\name{rss_txt_pull}
\alias{rss_txt_pull}
\title{Miscellaneous Codes for collecting text from RSS feeds}
\usage{
rss_txt_pull(url, paper, section = "")
}
\arguments{
\item{url}{}
}
\description{

}
\examples{

}
